+++ trunk/rpm.c
@@ -1,5 +1,5 @@
 /*
- * "$Id: rpm.c,v 1.26 2001/03/20 15:23:21 mike Exp $"
+ * "$Id: rpm.c,v 1.27 2001/03/26 16:18:47 mike Exp $"
  *
  *   Red Hat package gateway for the ESP Package Manager (EPM).
  *
@@ -364,18 +364,21 @@ make_rpm(const char     *prodname,	/* I - Product short name */
   * Remove temporary files...
   */
 
-  if (Verbosity)
-    puts("Removing temporary distribution files...");
+  if (!KeepFiles)
+  {
+    if (Verbosity)
+      puts("Removing temporary distribution files...");
 
-  sprintf(command, "/bin/rm -rf %s/buildroot", directory);
-  system(command);
+    sprintf(command, "/bin/rm -rf %s/buildroot", directory);
+    system(command);
 
-  unlink(specname);
+    unlink(specname);
+  }
 
   return (0);
 }
 
 
 /*
- * End of "$Id: rpm.c,v 1.26 2001/03/20 15:23:21 mike Exp $".
+ * End of "$Id: rpm.c,v 1.27 2001/03/26 16:18:47 mike Exp $".
  */
